zu #8721

+++ trunk/webEdition/we/include/we_classes/contents/we_document.class.php
@@ -268,9 +268,14 @@ class we_document extends we_root{
 	}
 
 	function formNavigation(){
-		$delallbut = we_html_button::create_button('delete_all', "javascript:if(confirm('" . g_l('navigation', '[dellall_question]') . "')) we_cmd('delete_all_navi')", true, 0, 0, "", "", (permissionhandler::hasPerm('EDIT_NAVIGATION') && $this->NavigationItems) ? false : true);
-		$addbut = we_html_button::create_button('add', "javascript:we_cmd('module_navigation_edit_navi',0)", true, 100, 22, '', '', (permissionhandler::hasPerm('EDIT_NAVIGATION') && $this->ID && $this->Published) ? false : true, false);
+		$isSee = $_SESSION['weS']['we_mode'] == we_base_constants::MODE_SEE;
 
+		if(permissionhandler::hasPerm('EDIT_NAVIGATION') && $isSee){
+			$delallbut = we_html_button::create_button('delete_all', "javascript:if(confirm('" . g_l('navigation', '[dellall_question]') . "')) we_cmd('delete_all_navi')", true, 0, 0, "", "", (permissionhandler::hasPerm('EDIT_NAVIGATION') && $this->NavigationItems) ? false : true);
+			$addbut = we_html_button::create_button('add', "javascript:we_cmd('module_navigation_edit_navi',0)", true, 100, 22, '', '', (permissionhandler::hasPerm('EDIT_NAVIGATION') && $this->ID && $this->Published) ? false : true, false);
+		} else {
+			$addbut = $delallbut = '';
+		}
 		$navis = new MultiFileChooser(508, $this->NavigationItems, 'delete_navi', we_html_button::create_button_table(array($delallbut, $addbut)), "module_navigation_edit_navi", "Icon,Path", NAVIGATION_TABLE);
 		$navis->extraDelFn = 'setScrollTo();';
 		$NoDelNavis = makeArrayFromCSV($this->NavigationItems);
@@ -285,9 +290,9 @@ class we_document extends we_root{
 				}
 			}
 		}
-		$navis->setDisabledDelItems(makeCSVFromArray($NoDelNavis), g_l('navigation', '[NoDeleteFromDocument]'));
+		$navis->setDisabledDelItems($NoDelNavis, g_l('navigation', '[NoDeleteFromDocument]'));
 
-		if(!permissionhandler::hasPerm('EDIT_NAVIGATION')){
+		if(!$isSee || !permissionhandler::hasPerm('EDIT_NAVIGATION')){
 			$navis->isEditable = false;
 			$navis->CanDelete = false;
 		}
